requirements.txt
flask
Jinja2
gunicorn
pylint
requests
pyyaml
python-dotenv
SQLAlchemy
mongoengine
Pillow
PyJWT
Flask-Session